+++ airframe/src/main/scala/wvlet/airframe/Binder.scala
@@ -17,6 +17,7 @@ import java.util.UUID
 
 import wvlet.airframe.AirframeException.CYCLIC_DEPENDENCY
 import wvlet.airframe.AirframeMacros._
+import wvlet.airframe.lifecycle.{BEFORE_SHUTDOWN, ON_INIT, ON_INJECT, ON_SHUTDOWN, ON_START}
 import wvlet.airframe.surface.Surface
 import wvlet.log.LogSupport