+++ airframe/src/test/scala/wvlet/airframe/lifecycle/CloseableShutdownHookTest.scala
@@ -11,10 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
 import java.util.concurrent.atomic.AtomicBoolean
 
+import wvlet.airframe.{bind, newSilentDesign}
 import wvlet.airspec.AirSpec
 
 /**