+++ airframe/src/test/scala/wvlet/airframe/lifecycle/LifeCycleManagerTest.scala
@@ -11,14 +11,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
-import java.io.Closeable
 import java.util.concurrent.atomic.AtomicInteger
 
 import wvlet.airframe.AirframeException.{MULTIPLE_SHUTDOWN_FAILURES, SHUTDOWN_FAILURE}
+import wvlet.airframe.{bind, newDesign, newSilentDesign}
 import wvlet.airspec.AirSpec
 import wvlet.log.{LogLevel, LogSupport, Logger}
+import wvlet.airframe._
 
 class Counter extends LogSupport {
   val initialized = new AtomicInteger(0)
@@ -244,41 +245,4 @@ class LifeCycleManagerTest extends AirSpec {
     e.causes.find(_.getMessage.contains("failure test")) shouldBe defined
     e.causes.find(_.getMessage.contains("failure 2")) shouldBe defined
   }
-
-  class LifeCycleTraitTest extends LifeCycle {
-    var onInitIsCalled: Boolean         = false
-    var onInjectIsCalled: Boolean       = false
-    var onStartIsCalled: Boolean        = false
-    var beforeShutdownIsCalled: Boolean = false
-    var onShutdownIsCalled: Boolean     = false
-    override def onInit: Unit = {
-      onInitIsCalled = true
-    }
-    override def onInject: Unit = {
-      onInjectIsCalled = true
-    }
-    override def onStart: Unit = {
-      onStartIsCalled = true
-    }
-    override def beforeShutdown: Unit = {
-      beforeShutdownIsCalled = true
-    }
-    override def onShutdown: Unit = {
-      onShutdownIsCalled = true
-    }
-  }
-
-  def `support LifeCycle trait`: Unit = {
-    val t = newSilentDesign
-      .build[LifeCycleTraitTest] { t =>
-        t
-      }.asInstanceOf[LifeCycleTraitTest]
-
-    t.onInitIsCalled shouldBe true
-    t.onInjectIsCalled shouldBe true
-    t.onStartIsCalled shouldBe true
-    t.beforeShutdownIsCalled shouldBe true
-    t.onShutdownIsCalled shouldBe true
-  }
-
 }